requirements.txt
Flask
Flask-Cors
celery
redis
langchain
langchain-openai
google-cloud-storage
requests
python-dotenv
behave
